Buttons for signup and logout testing to change userAuth working
