+++ client/src/App.css
@@ -44,11 +44,11 @@
   color: white;
   font-size: 16px;
   font-weight: bold;
+  transition: 0.5s;
 }
 
 .signup-form-button:hover {
   background-color: var(--hover-color1);
-  transition: 0.5s;
 }
 /* ----------- SIGN UP PAGE ----------*/
 
@@ -83,3 +83,7 @@ h2 {
   margin-left: 20px;
 }
 
+/* nav-bar */
+.nav-bar-links {
+  color: black;
+}
