+++ client/src/component/Nav.jsx
@@ -1,20 +1,25 @@
 import { Link } from "react-router-dom";
 
-export default function Nav() {
+export default function Nav(props) {
   return (
     <div>
       <nav>
-        <Link style={{ textDecoration: "none", color: "white" }} to="/home">
-          <span>Home</span>
+        <Link style={{ textDecoration: "none", color: "white" }} to="/">
+          <span className="nav-bar-links">Home</span>
         </Link>
         <Link style={{ textDecoration: "none", color: "white" }} to="/history">
-          <span>History</span>
+          <span className="nav-bar-links">History</span>
         </Link>
         <Link style={{ textDecoration: "none", color: "white" }} to="/summary">
-          <span>Summary</span>
+          <span className="nav-bar-links">Summary</span>
+        </Link>
+        <Link style={{ textDecoration: "none", color: "white" }} to="/about">
+          <span className="nav-bar-links">About</span>
         </Link>
         <Link style={{ textDecoration: "none", color: "white" }} to="/">
-          <span>Log Out</span>
+          <span onClick={props.authClick} className="nav-bar-links">
+            Log Out
+          </span>
         </Link>
       </nav>
     </div>
